swarm/storage: rename memstore_lrucache to memstore
